compress inmp to 16-bit stereo

+++ digital_1b/lib/I2S/I2S.h
@@ -7,7 +7,7 @@
 void i2s_init(i2s_chan_handle_t* input_chan_ptr, i2s_chan_handle_t* output_chan_ptr, uint32_t sample_rate, uint32_t frame_size, uint32_t dma_buffer_count);
 
 // writes frame_size samples to the DMA output buffer
-esp_err_t i2s_write_once(i2s_chan_handle_t* chan_handle_ptr, int32_t* data, size_t frame_size);
+esp_err_t i2s_write_once(i2s_chan_handle_t* chan_handle_ptr, int16_t* data, size_t frame_size);
 
 // reads frame_size samples from DMA over i2s
 esp_err_t i2s_read_once(i2s_chan_handle_t* chan_handle_ptr, int32_t* data, size_t frame_size);
